Move TableCard actions from packaged component to ReportTable.
