+++ client/analytics/components/report-table/ie.scss
@@ -1,6 +1,6 @@
 /** @format */
 
-.woocommerce-table {
+.woocommerce-report-table {
 	&.has-compare,
 	&.has-search {
 		.woocommerce-card__action {
