+++ client/ie.scss
@@ -1,0 +1,7 @@
+/**
+ * Internal Dependencies
+ *
+ * @format
+ */
+
+@import 'analytics/components/report-table/ie.scss';
