+++ packages/components/src/ie.scss
@@ -3,4 +3,3 @@
  */
 @import 'segmented-selection/ie.scss';
 @import 'summary/ie.scss';
-@import 'table/ie.scss';
