+++ packages/components/src/table/index.js
@@ -5,31 +5,17 @@
 import { __ } from '@wordpress/i18n';
 import classnames from 'classnames';
 import { Component, Fragment } from '@wordpress/element';
-import { find, first, isEqual, noop, partial, uniq, without } from 'lodash';
-import { IconButton } from '@wordpress/components';
+import { find, first, isEqual, noop, without } from 'lodash';
 import PropTypes from 'prop-types';
-
-/**
- * WooCommerce dependencies
- */
-import {
-	downloadCSVFile,
-	generateCSVDataFromTable,
-	generateCSVFileName,
-} from '@woocommerce/csv-export';
-import { getIdsFromQuery, getSearchWords, updateQueryString } from '@woocommerce/navigation';
 
 /**
  * Internal dependencies
  */
 import Card from '../card';
-import CompareButton from '../compare-filter/button';
-import DownloadIcon from './download-icon';
 import EllipsisMenu from '../ellipsis-menu';
 import MenuItem from '../ellipsis-menu/menu-item';
 import MenuTitle from '../ellipsis-menu/menu-title';
 import Pagination from '../pagination';
-import Search from '../search';
 import Table from './table';
 import TablePlaceholder from './placeholder';
 import TableSummary from './summary';
@@ -45,35 +31,17 @@ import TableSummary from './summary';
 class TableCard extends Component {
 	constructor( props ) {
 		super( props );
-		const { query, compareBy } = this.props;
 		const showCols = this.getShowCols( props.headers );
-		const selectedRows = query.filter ? getIdsFromQuery( query[ compareBy ] ) : [];
 
-		this.state = { showCols, selectedRows };
+		this.state = { showCols };
 		this.onColumnToggle = this.onColumnToggle.bind( this );
-		this.onClickDownload = this.onClickDownload.bind( this );
-		this.onCompare = this.onCompare.bind( this );
 		this.onPageChange = this.onPageChange.bind( this );
-		this.onSearchChange = this.onSearchChange.bind( this );
-		this.selectRow = this.selectRow.bind( this );
-		this.selectAllRows = this.selectAllRows.bind( this );
 	}
 
 	componentDidUpdate( { headers: prevHeaders, query: prevQuery } ) {
-		const { compareBy, headers, onColumnsChange, query } = this.props;
+		const { headers, onColumnsChange, query } = this.props;
 		const { showCols } = this.state;
 
-		if ( query.filter || prevQuery.filter ) {
-			const prevIds = prevQuery.filter ? getIdsFromQuery( prevQuery[ compareBy ] ) : [];
-			const currentIds = query.filter ? getIdsFromQuery( query[ compareBy ] ) : [];
-			if ( ! isEqual( prevIds.sort(), currentIds.sort() ) ) {
-				/* eslint-disable react/no-did-update-set-state */
-				this.setState( {
-					selectedRows: currentIds,
-				} );
-				/* eslint-enable react/no-did-update-set-state */
-			}
-		}
 		if ( ! isEqual( headers, prevHeaders ) ) {
 			/* eslint-disable react/no-did-update-set-state */
 			this.setState( {
@@ -147,36 +115,6 @@ class TableCard extends Component {
 		};
 	}
 
-	onClickDownload() {
-		const { query, onClickDownload, searchBy, title } = this.props;
-		const params = Object.assign( {}, query );
-
-		// Delete unnecessary items from filename.
-		delete params.extended_info;
-		if ( params.search ) {
-			delete params[ searchBy ];
-		}
-
-		// @todo The current implementation only downloads the contents displayed in the table.
-		// Another solution is required when the data set is larger (see #311).
-		downloadCSVFile(
-			generateCSVFileName( title, params ),
-			generateCSVDataFromTable( this.getVisibleHeaders(), this.getVisibleRows() )
-		);
-
-		if ( onClickDownload ) {
-			onClickDownload();
-		}
-	}
-
-	onCompare() {
-		const { compareBy, compareParam, onQueryChange } = this.props;
-		const { selectedRows } = this.state;
-		if ( compareBy ) {
-			onQueryChange( 'compare' )( compareBy, compareParam, selectedRows.join( ',' ) );
-		}
-	}
-
 	onPageChange( ...params ) {
 		const { onPageChange, onQueryChange } = this.props;
 		if ( onPageChange ) {
@@ -187,179 +125,32 @@ class TableCard extends Component {
 		}
 	}
 
-	onSearchChange( values ) {
-		const {
-			baseSearchQuery,
-			compareParam,
-			onSearch,
-			searchBy,
-		} = this.props;
-		// A comma is used as a separator between search terms, so we want to escape
-		// any comma they contain.
-		const labels = values.map( v => v.label.replace( ',', '%2C' ) );
-		if ( labels.length ) {
-			updateQueryString( {
-				filter: undefined,
-				[ compareParam ]: undefined,
-				[ searchBy ]: undefined,
-				...baseSearchQuery,
-				search: uniq( labels ).join( ',' ),
-			} );
-		} else {
-			updateQueryString( {
-				search: undefined,
-			} );
-		}
-
-		onSearch( values );
-	}
-
-	selectAllRows( event ) {
-		const { ids } = this.props;
-		if ( event.target.checked ) {
-			this.setState( {
-				selectedRows: ids,
-			} );
-		} else {
-			this.setState( {
-				selectedRows: [],
-			} );
-		}
-	}
-
-	selectRow( i, event ) {
-		const { ids } = this.props;
-		if ( event.target.checked ) {
-			this.setState( ( { selectedRows } ) => ( {
-				selectedRows: uniq( [ ids[ i ], ...selectedRows ] ),
-			} ) );
-		} else {
-			this.setState( ( { selectedRows } ) => {
-				const index = selectedRows.indexOf( ids[ i ] );
-				return {
-					selectedRows: [ ...selectedRows.slice( 0, index ), ...selectedRows.slice( index + 1 ) ],
-				};
-			} );
-		}
-	}
-
-	getCheckbox( i ) {
-		const { ids = [] } = this.props;
-		const { selectedRows } = this.state;
-		const isChecked = -1 !== selectedRows.indexOf( ids[ i ] );
-		return {
-			display: (
-				<input type="checkbox" onChange={ partial( this.selectRow, i ) } checked={ isChecked } />
-			),
-			value: false,
-		};
-	}
-
-	getAllCheckbox() {
-		const { ids = [] } = this.props;
-		const { selectedRows } = this.state;
-		const hasData = ids.length > 0;
-		const isAllChecked = hasData && ids.length === selectedRows.length;
-		return {
-			cellClassName: 'is-checkbox-column',
-			label: (
-				<input
-					type="checkbox"
-					onChange={ this.selectAllRows }
-					aria-label={ __( 'Select All' ) }
-					checked={ isAllChecked }
-					disabled={ ! hasData }
-				/>
-			),
-			required: true,
-		};
-	}
-
 	render() {
 		const {
-			compareBy,
-			downloadable,
-			labels = {},
+			actions,
+			className,
 			isLoading,
-			onClickDownload,
 			onQueryChange,
 			onSort,
 			query,
 			rowHeader,
 			rowsPerPage,
-			searchBy,
 			showMenu,
 			summary,
 			title,
 			totalRows,
 		} = this.props;
-		const { selectedRows, showCols } = this.state;
-		const searchWords = getSearchWords( query );
-		const searchedLabels = searchWords.map( v => ( { id: v, label: v } ) );
+		const { showCols } = this.state;
 		const allHeaders = this.props.headers;
-		let headers = this.getVisibleHeaders();
-		let rows = this.getVisibleRows();
-		if ( compareBy ) {
-			rows = rows.map( ( row, i ) => {
-				return [ this.getCheckbox( i ), ...row ];
-			} );
-			headers = [ this.getAllCheckbox(), ...headers ];
-		}
-		const hasData = 0 < totalRows;
-
-		const className = classnames( 'woocommerce-analytics__card', {
-			'woocommerce-table': true,
-			'has-compare': !! compareBy,
-			'has-search': !! searchBy,
-		} );
+		const headers = this.getVisibleHeaders();
+		const rows = this.getVisibleRows();
+		const classes = classnames( 'woocommerce-table', className );
 
 		return (
 			<Card
-				className={ className }
+				className={ classes }
 				title={ title }
-				action={ [
-					compareBy && (
-						<CompareButton
-							key="compare"
-							className="woocommerce-table__compare"
-							count={ selectedRows.length }
-							helpText={
-								labels.helpText || __( 'Check at least two items below to compare', 'woocommerce-admin' )
-							}
-							onClick={ this.onCompare }
-							disabled={ ! hasData }
-						>
-							{ labels.compareButton || __( 'Compare', 'woocommerce-admin' ) }
-						</CompareButton>
-					),
-					searchBy && (
-						<Search
-							allowFreeTextSearch={ true }
-							inlineTags
-							key="search"
-							onChange={ this.onSearchChange }
-							placeholder={ labels.placeholder || __( 'Search by item name', 'woocommerce-admin' ) }
-							selected={ searchedLabels }
-							showClearButton={ true }
-							type={ searchBy }
-							disabled={ ! hasData }
-						/>
-					),
-					( downloadable || onClickDownload ) && (
-						<IconButton
-							key="download"
-							className="woocommerce-table__download-button"
-							disabled={ isLoading || ! hasData }
-							onClick={ this.onClickDownload }
-							isLink
-						>
-							<DownloadIcon />
-							<span className="woocommerce-table__download-button__label">
-								{ labels.downloadButton || __( 'Download', 'woocommerce-admin' ) }
-							</span>
-						</IconButton>
-					),
-				] }
+				action={ actions }
 				menu={
 					showMenu && <EllipsisMenu label={ __( 'Choose which values to display', 'woocommerce-admin' ) }
 						renderContent={ () => (
@@ -427,14 +218,6 @@ class TableCard extends Component {
 
 TableCard.propTypes = {
 	/**
-	 * The string to use as a query parameter when comparing row items.
-	 */
-	compareBy: PropTypes.string,
-	/**
-	 * Url query parameter compare function operates on
-	 */
-	compareParam: PropTypes.string,
-	/**
 	 * An array of column headers (see `Table` props).
 	 */
 	headers: PropTypes.arrayOf(
@@ -443,19 +226,10 @@ TableCard.propTypes = {
 			defaultSort: PropTypes.bool,
 			isSortable: PropTypes.bool,
 			key: PropTypes.string,
-			label: PropTypes.string,
+			label: PropTypes.oneOfType( [ PropTypes.string, PropTypes.node ] ),
 			required: PropTypes.bool,
 		} )
 	),
-	/**
-	 * Custom labels for table header actions.
-	 */
-	labels: PropTypes.shape( {
-		compareButton: PropTypes.string,
-		downloadButton: PropTypes.string,
-		helpText: PropTypes.string,
-		placeholder: PropTypes.string,
-	} ),
 	/**
 	 * A list of IDs, matching to the row list so that ids[ 0 ] contains the object ID for the object displayed in row[ 0 ].
 	 */
@@ -474,21 +248,9 @@ TableCard.propTypes = {
 	 */
 	onColumnsChange: PropTypes.func,
 	/**
-	 * A function which is called upon the user searching in the table header.
-	 */
-	onSearch: PropTypes.func,
-	/**
 	 * A function which is called upon the user changing the sorting of the table.
 	 */
 	onSort: PropTypes.func,
-	/**
-	 * Whether the table must be downloadable. If true, the download button will appear.
-	 */
-	downloadable: PropTypes.bool,
-	/**
-	 * A callback function called when the "download" button is pressed. Optional, if used, the download button will appear.
-	 */
-	onClickDownload: PropTypes.func,
 	/**
 	 *  An object of the query parameters passed to the page, ex `{ page: 2, per_page: 5 }`.
 	 */
@@ -513,10 +275,6 @@ TableCard.propTypes = {
 	 */
 	rowsPerPage: PropTypes.number.isRequired,
 	/**
-	 * The string to use as a query parameter when searching row items.
-	 */
-	searchBy: PropTypes.string,
-	/**
 	 * Boolean to determine whether or not ellipsis menu is shown.
 	 */
 	showMenu: PropTypes.bool,
@@ -538,25 +296,17 @@ TableCard.propTypes = {
 	 * The total number of rows (across all pages).
 	 */
 	totalRows: PropTypes.number.isRequired,
-	/**
-	 * Pass in query parameters to be included in the path when onSearch creates a new url.
-	 */
-	baseSearchQuery: PropTypes.object,
 };
 
 TableCard.defaultProps = {
-	compareParam: 'filter',
-	downloadable: false,
 	isLoading: false,
 	onQueryChange: noop,
 	onColumnsChange: noop,
-	onSearch: noop,
 	onSort: undefined,
 	query: {},
 	rowHeader: 0,
 	rows: [],
 	showMenu: true,
-	baseSearchQuery: {},
 };
 
 export default TableCard;
